Refactor simulation into modules
