+++ AGENTS.md
@@ -1,0 +1,4 @@
+# Guidelines
+
+- Lint the code with `npm run lint` and run tests with `npm test` whenever source code inside `src/` changes.
+- Tests and lint are not required for documentation-only edits.
