+++ eslint.config.js
@@ -1,0 +1,14 @@
+export default [
+  {
+    files: ['**/*.js'],
+    languageOptions: {
+      ecmaVersion: 'latest',
+      sourceType: 'module',
+      globals: {
+        window: 'readonly',
+        document: 'readonly'
+      }
+    },
+    rules: {}
+  }
+];
