+++ index.html
@@ -20,7 +20,7 @@
     <div id="vis"></div>
 
     <!-- Ton script principal en module (importe Three.js, Cannon‑ES, bird‑oid/Boids, etc.) -->
-    <script type="module" src="simulation.js"></script>
+    <script type="module" src="src/main.js"></script>
 </body>
 
 </html>
