+++ package.json
@@ -1,0 +1,17 @@
+{
+  "name": "la_bonne_echappee",
+  "version": "1.0.0",
+  "description": "",
+  "main": "simulation.js",
+  "scripts": {
+    "test": "node -e \"console.log('Tests executed')\"",
+    "lint": "eslint ."
+  },
+  "keywords": [],
+  "author": "",
+  "license": "ISC",
+  "type": "module",
+  "devDependencies": {
+    "eslint": "^9.31.0"
+  }
+}
