+++ src/main.js
@@ -1,0 +1,9 @@
+import './setupScene.js';
+import './physicsWorld.js';
+import './track.js';
+import './riders.js';
+import './ui.js';
+import './startButton.js';
+import { animate } from './animation.js';
+
+animate();
