+++ src/physicsWorld.js
@@ -1,0 +1,24 @@
+import * as CANNON from 'https://cdn.jsdelivr.net/npm/cannon-es@0.20.0/dist/cannon-es.js';
+
+const world = new CANNON.World();
+world.gravity.set(0, 0, 0);
+world.broadphase = new CANNON.SAPBroadphase(world);
+world.solver.iterations = 10;
+
+const defaultMaterial = new CANNON.Material('defaultMaterial');
+const contactMaterial = new CANNON.ContactMaterial(defaultMaterial, defaultMaterial, { friction: 0.3, restitution: 0.1 });
+world.addContactMaterial(contactMaterial);
+world.defaultContactMaterial = contactMaterial;
+
+let physicsAccumulator = 0;
+const fixedTimeStep = 1 / 60;
+
+function stepPhysics(dt) {
+  physicsAccumulator += dt;
+  while (physicsAccumulator >= fixedTimeStep) {
+    world.step(fixedTimeStep);
+    physicsAccumulator -= fixedTimeStep;
+  }
+}
+
+export { CANNON, world, stepPhysics };
